+++ astropy/modeling/rotations.py
@@ -7,9 +7,12 @@ Implements rotations, including spherical rotations as defined in WCS Paper II
 `RotateNative2Celestial` and `RotateCelestial2Native` follow the convention in
 WCS Paper II to rotate to/from a native sphere and the celestial sphere.
 
-The user interface sets and displays angles in degrees but the values are
-stored internally in radians.  This is managed through the parameter
-setters/getters.
+The implementation uses `EulerAngleRotation`. The model parameters are
+three angles: the longitude (``lon``) and latitude (``lat``) of the fiducial point
+in the celestial system (``CRVAL`` keywords in FITS), and the longitude of the celestial
+pole in the native system (``lon_pole``). The Euler angles are ``lon+90``, ``90 - lat``
+and ``-(lon_pole - 90)``.
+
 
 References
 ----------
@@ -55,7 +58,7 @@ class EulerAngleRotation(Model):
     theta = Parameter(default=0, getter=np.rad2deg, setter=np.deg2rad)
     psi = Parameter(default=0, getter=np.rad2deg, setter=np.deg2rad)
 
-    def __init__(self, phi, theta, psi, axes_order):
+    def __init__(self, phi, theta, psi, axes_order, **kwargs):
         self.axes = ['x', 'y', 'z']
         if len(axes_order) != 3:
             raise TypeError(
@@ -66,7 +69,7 @@ class EulerAngleRotation(Model):
             raise ValueError("Unrecognized axis label {0}; "
                              "should be one of {1} ".format(unrecognized, self.axes))
         self.axes_order = axes_order
-        super(EulerAngleRotation, self).__init__(phi=phi, theta=theta, psi=psi)
+        super(EulerAngleRotation, self).__init__(phi=phi, theta=theta, psi=psi, **kwargs)
 
     def _create_matrix(self, phi, theta, psi, axes_order):
         matrices = []
@@ -112,6 +115,13 @@ class EulerAngleRotation(Model):
         z = np.sin(delta)
         return np.array([x, y, z])
 
+    @staticmethod
+    def cartesian2spherical(x, y, z):
+        h = np.hypot(x, y)
+        alpha  = np.rad2deg(np.arctan2(y, x))
+        delta = np.rad2deg(np.arctan2(z, h))
+        return alpha, delta
+
     def inverse(self):
         return self.__class__(phi=-self.psi,
                               theta=-self.theta,
@@ -119,54 +129,37 @@ class EulerAngleRotation(Model):
                               axes_order=self.axes_order[::-1])
 
     def evaluate(self, alpha, delta, phi, theta, psi):
+        shape = None
+        if isinstance(alpha, np.ndarray) and alpha.ndim == 2:
+            alpha = alpha.flatten()
+            delta = delta.flatten()
+            shape = alpha.shape
         inp = self.spherical2cartesian(alpha, delta)
         matrix = self._create_matrix(phi, theta, psi, self.axes_order)
         result = np.dot(matrix, inp)
-        return (np.rad2deg(np.arctan2(result[1], result[0])),
-                np.rad2deg(np.arcsin(result[2])))
+        a, b = self.cartesian2spherical(*result)
+        if shape is not None:
+            a.shape = shape
+            b.shape = shape
+        return a, b
 
 
-class _SkyRotation(Model):
+class _SkyRotation(EulerAngleRotation):
     """
-    Base class for FITS WCS sky rotations.
-
-    Parameters
-    ----------
-    lon : float
-        Celestial longitude of the fiducial point.
-    lat : float
-        Celestial latitude of the fiducial point.
-    lon_pole : float
-        Longitude of the celestial pole in the native system.
+    Base class for RotateNative2Celestial and RotateCelestial2Native.
     """
 
-    lon = Parameter(default=0, getter=np.rad2deg, setter=np.deg2rad)
-    lat = Parameter(default=0, getter=np.rad2deg, setter=np.deg2rad)
-    lon_pole = Parameter(default=0, getter=np.rad2deg, setter=np.deg2rad)
+    def __init__(self, phi, theta, psi, **kwargs):
+        super(_SkyRotation, self).__init__(phi, theta, psi, 'zxz', **kwargs)
 
-    @staticmethod
-    def _rotate_zxz(phi_i, theta_i, lon, lat, lon_pole):
-        """
-        Defines a ZXZ rotation from initial coordinates phi_i, theta_i.
-
-        All inputs and outputs are in radians.
-        """
-
-        cos_theta_i = np.cos(theta_i)
-        sin_theta_i = np.sin(theta_i)
-        cos_lat = np.cos(lat)
-        sin_lat = np.sin(lat)
-        delta = phi_i - lon_pole
-        cos_delta = np.cos(delta)
-
-        phi_f = lon + np.arctan2(-cos_theta_i * np.sin(delta),
-                                 sin_theta_i * cos_lat -
-                                 cos_theta_i * sin_lat * cos_delta)
-
-        theta_f = np.arcsin(sin_theta_i * sin_lat +
-                            cos_theta_i * cos_lat * cos_delta)
-
-        return phi_f, theta_f
+    def evaluate(self, phi, theta, lon, lat, lon_pole):
+        alpha, delta = super(_SkyRotation, self).evaluate(phi, theta, lon, lat, lon_pole)
+        mask = alpha < 0
+        if isinstance(mask, np.ndarray):
+            alpha[mask] +=360
+        else:
+            alpha +=360
+        return alpha, delta
 
 
 class RotateNative2Celestial(_SkyRotation):
@@ -183,39 +176,53 @@ class RotateNative2Celestial(_SkyRotation):
         Longitude of the celestial pole in the native system.
     """
 
-    inputs = ('phi_N', 'theta_N')
-    outputs = ('alpha_C', 'delta_C')
+    # angles in degrees on the native sphere
+    inputs = ('phi', 'theta')
+    outputs = ('alpha', 'delta')
+
+    def __init__(self, lon, lat, lon_pole, **kwargs):
+        # Convert to Euler angles
+        phi = lon_pole - 90
+        theta = - (90 - lat)
+        psi = -(90 + lon)
+        super(RotateNative2Celestial, self).__init__(phi, theta, psi, **kwargs)
+
+    @property
+    def lon(self):
+        return - (self.psi + 90)
+
+    @lon.setter
+    def lon(self, val):
+        self.psi = - (val + 90)
+
+    @property
+    def lat(self):
+        return self.theta + 90
+
+    @lat.setter
+    def lat(self, val):
+        self.theta = val - 90
+
+    @property
+    def lon_pole(self):
+        return self.phi + 90
+
+    @lon_pole.setter
+    def lon_pole(self, val):
+        self.phi = val - 90
 
     @property
     def inverse(self):
-        return RotateCelestial2Native(self.lon, self.lat, self.lon_pole)
-
-    @classmethod
-    def evaluate(cls, phi_N, theta_N, lon, lat, lon_pole):
-        """
-        Rotate native spherical coordinates into celestial coordinates.
-        """
-
-        phi_N = np.deg2rad(phi_N)
-        theta_N = np.deg2rad(theta_N)
-
-        alpha_C, delta_C = cls._rotate_zxz(phi_N, theta_N, lon, lat, lon_pole)
-
-        alpha_C = np.rad2deg(alpha_C)
-        delta_C = np.rad2deg(delta_C)
-
-        mask = alpha_C < 0
-        if isinstance(mask, np.ndarray):
-            alpha_C[mask] += 360
-        elif mask:
-            alpha_C += 360
-
-        return alpha_C, delta_C
+        # convert to angles on the celestial sphere
+        lon = -(self.psi + 90)
+        lat = self.theta + 90
+        lon_pole = self.phi + 90
+        return RotateCelestial2Native(lon, lat, lon_pole)
 
 
 class RotateCelestial2Native(_SkyRotation):
     """
-    Transform from Celestial to Native to Spherical Coordinates.
+    Transform from Celestial to Native Spherical Coordinates.
 
     Parameters
     ----------
@@ -227,36 +234,50 @@ class RotateCelestial2Native(_SkyRotation):
         Longitude of the celestial pole in the native system.
     """
 
-    inputs = ('alpha_C', 'delta_C')
-    outputs = ('phi_N', 'theta_N')
+    # angles in degrees on the celestial sphere
+    inputs = ('alpha', 'delta')
+    # angles in degrees on the native sphere
+    outputs = ('phi', 'theta')
+
+
+    def __init__(self, lon, lat, lon_pole, **kwargs):
+        # Convert to Euler angles
+        phi = (90 + lon)
+        theta =  (90 - lat)
+        psi = -(lon_pole - 90)
+        super(RotateCelestial2Native, self).__init__(phi, theta, psi, **kwargs)
+
+    @property
+    def lon(self):
+        return self.phi - 90
+
+    @lon.setter
+    def lon(self, val):
+        self.phi = val + 90
+
+    @property
+    def lat(self):
+        return 90 - self.theta
+
+    @lat.setter
+    def lat(self, val):
+        self.theta = 90 - val
+
+    @property
+    def lon_pole(self):
+        return 90 - self.psi
+
+    @lon_pole.setter
+    def lon_pole(self, val):
+        self.psi = 90 - val
 
     @property
     def inverse(self):
-        return RotateNative2Celestial(self.lon, self.lat, self.lon_pole)
-
-    @classmethod
-    def evaluate(cls, alpha_C, delta_C, lon, lat, lon_pole):
-        """
-        Rotate celestial coordinates into native spherical coordinates.
-
-        This is the inverse transformation of RotateNative2Celestial.
-        """
-
-        alpha_C = np.deg2rad(alpha_C)
-        delta_C = np.deg2rad(delta_C)
-
-        phi_N, theta_N = cls._rotate_zxz(alpha_C, delta_C, lon_pole, lat, lon)
-
-        phi_N = np.rad2deg(phi_N)
-        theta_N = np.rad2deg(theta_N)
-
-        mask = phi_N > 180
-        if isinstance(mask, np.ndarray):
-            phi_N[mask] -= 360
-        elif mask:
-            phi_N -= 360
-
-        return phi_N, theta_N
+        # convert to angles on the celestial sphere
+        lon = -(self.psi + 90)
+        lat = self.theta + 90
+        lon_pole = self.phi + 90
+        return RotateCelestial2Native(lon, lat, lon_pole)
 
 
 class Rotation2D(Model):