+++ CHANGES.rst
@@ -217,6 +217,9 @@ API changes
   - Inputs (``coords`` and ``out``) to ``render`` function in ``Model`` are
     converted to float. [#4697]
 
+  - ``RotateNative2Celestial`` and ``RotateCelestial2Native`` are now
+    implemented as subclasses of ``EulerAngleRotation``. [#4881]
+
 - ``astropy.nddata``
 
   - ``NDDataBase`` does not set the private uncertainty property anymore. This